+++ src/app/static/src/app/localStorageManager.ts
@@ -50,6 +50,7 @@ export class LocalStorageManager {
             localStorage.clear();
             localStorage.setItem("user", currentUser);
         }
+
         const item = window.localStorage.getItem(appStateKey);
         if (item) {
             return JSON.parse(item) as Partial<RootState>;