+++ src/app/static/src/tests/components/files/manageFile.test.ts
@@ -1,30 +1,25 @@
-import Vue from "vue";
 import {shallowMount, Slots} from '@vue/test-utils';
 
 import ErrorAlert from "../../../app/components/ErrorAlert.vue";
 import Tick from "../../../app/components/Tick.vue";
 import FileUpload from "../../../app/components/files/FileUpload.vue";
 import ManageFile from "../../../app/components/files/ManageFile.vue";
-import {mockError, mockFile} from "../../mocks";
+import {mockDataExplorationState, mockError, mockFile} from "../../mocks";
 import LoadingSpinner from "../../../app/components/LoadingSpinner.vue";
 import Vuex, {Store} from "vuex";
-import {emptyState, RootState} from "../../../app/root";
 import registerTranslations from "../../../app/store/translations/registerTranslations";
-import {expectTranslated} from "../../testHelpers";
+import {expectTranslated, expectTranslatedWithStoreType} from "../../testHelpers";
+import {DataExplorationState, initialDataExplorationState} from "../../../app/store/dataExploration/dataExploration";
 
 describe("Manage file component", () => {
 
-    const mockGetters = {
-        editsRequireConfirmation: () => false
-    };
-
-    const createStore = () => {
+    const createStore = (customStore = initialDataExplorationState(), requireConfirmation = false) => {
         const store = new Vuex.Store({
-            state: emptyState(),
+            state: customStore,
             modules: {
                 stepper: {
                     namespaced: true,
-                    getters: mockGetters
+                    getters: {editsRequireConfirmation: () => requireConfirmation}
                 }
             }
         });
@@ -32,7 +27,7 @@ describe("Manage file component", () => {
         return store;
     };
 
-    const createSut = (props?: any, slots?: Slots, store?: Store<RootState>) => {
+    const createSut = (props?: any, slots?: Slots, store?: Store<DataExplorationState>) => {
         return shallowMount(ManageFile, {
             store: store || createStore(),
             propsData: {
@@ -87,7 +82,7 @@ describe("Manage file component", () => {
         const wrapper = createSut({
             existingFileName: "existing-name.csv"
         }, undefined, store);
-        expectTranslated(wrapper.find("label.file-name strong"), "File", "Fichier", "Ficheiro", store);
+        expectTranslatedWithStoreType(wrapper.find("label.file-name strong"), "File", "Fichier", "Ficheiro", store);
         expect(wrapper.find("label.file-name").text()).toContain("existing-name.csv");
     });
 
@@ -119,6 +114,31 @@ describe("Manage file component", () => {
             existingFileName: "File.csv",
             deleteFile: removeHandler
         });
+        const removeLink = wrapper.find("a");
+        expect(removeLink.text()).toBe("remove");
+
+        removeLink.trigger("click");
+
+        expect(removeHandler.mock.calls.length).toBe(1);
+    });
+
+    it("renders remove link if existing filename is present when on data exploration mode", () => {
+        const removeHandler = jest.fn();
+
+        const wrapper = shallowMount(ManageFile, {
+            store: createStore(mockDataExplorationState(), true),
+            propsData: {
+                error: null,
+                label: "PJNZ",
+                valid: true,
+                upload: jest.fn(),
+                deleteFile: removeHandler,
+                name: "pjnz",
+                accept: "csv",
+                existingFileName: "File.csv"
+            }
+        });
+
         const removeLink = wrapper.find("a");
         expect(removeLink.text()).toBe("remove");