+++ NEWS.md
@@ -1,6 +1,64 @@
-# hint 1.69.0
+# hint 1.74.2
 
 * Update review output plots to use metadata from modelCalibrate
+
+# hint 1.74.1
+
+* Remove calibration plot when user changes calibration options
+
+# hint 1.74.0
+
+* Add navigation buttons to bottom of every page
+
+# hint 1.73.2
+
+* When importing from ADR remove all datasets currently uploaded
+
+# hint 1.73.1
+
+* fix issues with displaying Input Time Series for DRC data
+
+# hint 1.73.0
+
+* prompt login when using data exploration mode
+
+# hint 1.72.3
+
+* Make time series tooltip usable
+
+# hint 1.72.1
+
+* Troubleshooting request very slow to close the dialog and send multiple errors
+
+# hint 1.72.0
+
+* Add DataExploration app component
+
+# hint 1.71.1
+
+* bug fix for input indicators being included when not relevant for dataset
+
+# hint 1.71.0
+
+* support redirecting to a requested url after a successful login
+
+# hint 1.70.3
+
+* Move error reporting related mutations out of root mutations
+
+# hint 1.70.2
+
+* bug fix for undefined indicator in choropleth while new data is loading
+
+# hint 1.70.1
+
+* correct Portuguese translation for "Steps to reproduce"
+
+# hint 1.70.0
+
+* adds basic data exploration app with loading spinner 
+
+# hint 1.69.0
 
 * Add feedback to error reporting